Add disqus shortname and disqusId configurations
This commit updates `_config.icarus.yml` by adding a `disqus shortname` of `'astroxu'`, facilitating Disqus comment integration. Additionally, a `disqusId` of `20241210` is added to the `2024_a_b.md` post to track comments for that specific post. Several whitespace inconsistencies were corrected to enhance the overall readability of the configuration file.